backend/requirements.txt
#==========================================
#IDENTITY: The Arsenal / Dependencies Matrix
#FILEPATH: backend/requirements.txt
#COMPONENT: Backend Environment
#ROLE: All the sauce needed to make Python go brrrr.
#VIBE: Installing these is like funding your broker account. Necessary for the W.
#==========================================

#Core Framework - Fast and Furious

fastapi==0.104.1
uvicorn[standard]==0.24.0

#Database & ORM - The Vault

sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

#Real-time Comms - The Blast Protocol

websockets==12.0

#Security & Data Validation

pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0 # For that AES-256 vibe
passlib[bcrypt]==1.7.4

#The Agents & AI

google-generativeai==0.3.1 # Gemini 2.5 Flash for Orbit-Speak

#Forex Guardian 📈

#MetaTrader5==5.0.45
pandas==2.1.3

#Redis & Celery for Background Sync

redis==5.0.1
